insights/requirements.txt
fastapi>=0.68.0
uvicorn[standard]>=0.15.0
dash>=2.0.0
dash-bootstrap-components>=1.0.0
pandas>=1.3.0
sqlalchemy>=1.4.0
requests>=2.26.0
plotly>=5.0.0
python-dotenv>=0.19.0
